requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
requests==2.31.0
pika==1.3.2
PyJWT==2.8.0
email-validator==2.1.1